updated controller and migration for lat and long

+++ db/migrate/20200212171001_create_destinations.rb
@@ -3,8 +3,8 @@ class CreateDestinations < ActiveRecord::Migration[5.2]
     create_table :destinations do |t|
       t.string :title, null: false
       t.string :description, null: false
-      t.integer :latitude, null: false
-      t.integer :longitude, null: false
+      t.float :latitude, null: false
+      t.float :longitude, null: false
       t.string :google_places_place_id, null: false
 
       t.belongs_to :user, null: false
